Move image header probing to Cloud fs
